Rename: 폴더명에 따른 페이지 정리
