+++ src/pages/CategoryPage/index.tsx
@@ -1,5 +1,5 @@
-const Pottery = () => {
+const CategoryPage = () => {
   // eslint-disable-next-line react/react-in-jsx-scope
   return <div>도자기 카테고리 입니다.</div>;
 };
-export default Pottery;
+export default CategoryPage;
