+++ src/pages/index.tsx
@@ -1,2 +1,9 @@
-export { default as MainPage } from './MainPage/MainPage';
+export { default as MainPage } from './MainPage';
 export { default as ProductsDetailPage } from './ProductsDetailPage';
+export { default as AuthorPage } from './AuthorPage';
+export { default as CategoryPage } from './CategoryPage';
+export { default as FeedPage } from './FeedPage';
+export { default as LoginPage } from './LoginPage';
+export { default as ReservationPage } from './ReservationPage';
+export { default as SearchPage } from './SearchPage';
+export { default as UserPage } from './UserPage';
